[SYCL][CUDA] Enable group algorithm tests
Moves isSupportedDevice into support.h and adds check for CUDA.
Increases work-group size for some tests to ensure more than one warp.
